# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/channels": {
      "get": {
        "description": "查询所有支付渠道配置（敏感字段已脱敏）",
        "operationId": "ChannelConfigController_listChannels",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回渠道列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "支付渠道列表",
        "tags": [
          "管理后台"
        ]
      },
      "post": {
        "operationId": "ChannelConfigController_createChannel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "渠道创建成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "创建支付渠道",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/channels/{code}": {
      "put": {
        "operationId": "ChannelConfigController_updateChannel",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "渠道更新成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "更新支付渠道",
        "tags": [
          "管理后台"
        ]
      },
      "delete": {
        "operationId": "ChannelConfigController_deleteChannel",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "删除支付渠道",
        "tags": [
          "管理后台"
        ]
      }
    },
    "/admin/channels/{code}/test": {
      "post": {
        "description": "验证渠道是否可用",
        "operationId": "ChannelConfigController_testChannel",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "渠道可用"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "测试支付渠道",
        "tags": [
          "管理后台"
        ]
      }
    },
...
    "/admin/reconciliation/channel-bill/checks": {
      "get": {
        "summary": "渠道账单对账检查结果",
        "description": "查询渠道账单逐笔核对结果列表。",
        "operationId": "ReconciliationController_listChannelBillChecks",
        "responses": {
          "200": {
            "description": "处理成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "tags": [
          "对账"
        ]
      }
    },
    "/admin/reconciliation/channel-bill/generate": {
      "post": {
        "summary": "生成渠道账单",
        "description": "触发从渠道拉取当日账单文件并落库。",
        "operationId": "ReconciliationController_generateChannelBill",
        "responses": {
          "200": {
            "description": "处理成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "tags": [
          "对账"
        ]
      }
    },
    "/admin/reconciliation/channel-bill/reconcile": {
      "post": {
        "summary": "渠道账单对账",
        "description": "对已落库的渠道账单与平台收单订单进行逐笔对账。",
        "operationId": "ReconciliationController_reconcileChannelBill",
        "responses": {
          "200": {
            "description": "处理成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "tags": [
          "对账"
        ]
      }
    },
    "/admin/reconciliation/reports": {
      "get": {
        "operationId": "ReconciliationController_getReports",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回对账报告列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "对账报告列表",
        "tags": [
          "财务"
        ]
      }
    },
    "/admin/reconciliation/reports/export": {
      "get": {
        "operationId": "ReconciliationController_exportReports",
        "parameters": [],
        "responses": {
          "200": {
            "description": "CSV 文件下载"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "导出对账报告 CSV",
        "tags": [
          "财务"
        ]
      }
    },
    "/admin/reconciliation/reports/{date}": {
      "get": {
        "operationId": "ReconciliationController_getReport",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回对账报告详情"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "查询指定日期对账报告",
        "tags": [
          "财务"
        ]
      }
    },
    "/admin/reconciliation/run": {
      "post": {
        "description": "对指定日期进行对账",
        "operationId": "ReconciliationController_runReconciliation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunReconciliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "对账完成"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "执行对账",
        "tags": [
          "财务"
        ]
      }
    },
...
    "/cashier/orders": {
      "post": {
        "description": "用户发起向商户的付款订单",
        "operationId": "CashierController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCashierOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "订单创建成功"
          },
          "400": {
            "description": "KB601 商户订单号已存在"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "创建收银台订单",
        "tags": [
          "收银台"
        ]
      },
      "get": {
        "description": "分页查询当前用户的收银台订单",
        "operationId": "CashierController_listMyOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回订单列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "查询我的收银台订单",
        "tags": [
          "收银台"
        ]
      }
    },
    "/cashier/orders/export": {
      "get": {
        "description": "导出当前用户的订单为 CSV 文件",
        "operationId": "CashierController_exportMyOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "CSV 文件下载"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "导出订单 CSV",
        "tags": [
          "收银台"
        ]
      }
    },
    "/cashier/orders/reconciliation": {
      "get": {
        "description": "查询指定日期范围的对账数据",
        "operationId": "CashierController_reconciliation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回对账数据"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "对账查询",
        "tags": [
          "收银台"
        ]
      }
    },
    "/cashier/orders/{orderNo}": {
      "get": {
        "operationId": "CashierController_getOrder",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回订单详情"
          },
          "404": {
            "description": "KB603 订单不存在"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "查询订单详情",
        "tags": [
          "收银台"
        ]
      }
    },
    "/cashier/orders/{orderNo}/channel-pay": {
      "post": {
        "summary": "发起渠道支付",
        "description": "由收银台前端调用，向所选渠道发起支付并返回支付凭证（扫码链接/跳转参数）。",
        "operationId": "CashierController_channelPay",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "处理成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "tags": [
          "收银台"
        ]
      }
    },
    "/cashier/orders/{orderNo}/notify": {
      "post": {
        "description": "手动触发商户回调通知",
        "operationId": "CashierController_retryNotify",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "通知已发送"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "重试回调通知",
        "tags": [
          "收银台"
        ]
      }
    },
    "/cashier/qrcode/{code}": {
      "get": {
        "description": "扫描商户收款码获取订单信息（无需登录）",
        "operationId": "CashierQrCodeController_getQrCodeOrderInfo",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回收款码关联的订单信息"
          },
          "404": {
            "description": "KB610 收款码无效"
          }
        },
        "summary": "扫码获取收款信息",
        "tags": [
          "收银台"
        ]
      }
    },
...
  },
  "info": {
    "title": "科佰支付 KeBaiPay API",
    "description": "合规聚合支付技术服务商接口文档：资金绝不过平台，仅做收单、退款与对账聚合。\n\n## 认证方式\n- **用户接口**：使用 JWT Bearer Token 认证\n- **商户开放 API**：使用 HMAC-SHA256 签名认证\n- **管理后台**：使用管理员 JWT Token 认证\n\n## 错误码规范\n所有错误返回统一格式：`{ \"statusCode\": number, \"message\": \"KBxxx 错误描述\" }`\n\n## 金额说明\n数据库存储单位为「分」，接口传输单位为「元」\n\n> 已下线接口（v0.3.x 精简）：多平台对账聚合（/admin/channel-reconciliation/*）、AI 风控审计（/risk-audit/*、/admin/risk-audit/*）、自定义风控规则（/admin/risk-rules/custom、/risk-rules/custom）。对账以 /admin/reconciliation 为准。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "user-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "用户/管理员 JWT Token"
      },
      "app-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-App-Id",
        "description": "商户应用 App ID"
      }
    },
    "schemas": {
...
      "RunReconciliationDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCashierOrderDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateChannelConfigDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateChannelConfigDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}